+++ src/test/java/com/exceptionless/exceptionlessclient/configuration/PluginManagerTest.java
@@ -2,6 +2,7 @@ package com.exceptionless.exceptionlessclient.configuration;
 
 import com.exceptionless.exceptionlessclient.models.EventPluginContext;
 import com.exceptionless.exceptionlessclient.plugins.EventPluginIF;
+import com.exceptionless.exceptionlessclient.plugins.PluginManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
